chore(config): assign uint8_t

+++ main/ConfigManager.cpp
@@ -538,8 +538,8 @@ void ConfigManager::deserialize(msgpack_object obj, std::string type) {
                 std::ranges::for_each(msgpack_elements, [&](const msgpack_object& o) {
                     if (o.type == MSGPACK_OBJECT_ARRAY && o.via.array.size >= 2) {
                         const msgpack_object* inner_array_ptr = o.via.array.ptr;
-                        uint64_t key_val = inner_array_ptr[0].via.u64;
-                        uint64_t value_val = inner_array_ptr[1].via.u64;
+                        uint8_t key_val = inner_array_ptr[0].via.u64;
+                        uint8_t value_val = inner_array_ptr[1].via.u64;
                         (*arg)[static_cast<espConfig::actions_config_t::colorMap>(key_val)] = static_cast<uint8_t>(value_val);
                     }
                 });
